\name{SurvSBF-package}
\alias{SurvSBF-package}
\alias{SurvSBF}
\docType{package}
\title{
\packageTitle{SurvSBF}
}
\description{
\packageDescription{SurvSBF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SurvSBF}
\packageIndices{SurvSBF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SurvSBF}

Maintainer: \packageMaintainer{SurvSBF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
